authenticate and authorize with report

+++ books/controllers/reportsController.js
@@ -1,57 +1,76 @@
 const Report = require('../models/report');
+const { validationResult } = require('express-validator');
 
-exports.getAllReports = async (req, res) => {
+module.exports.getAllReports = async (req, res) => {
     try {
-        const reports = await Report.find();
-        res.json(reports);
+        const reports = await Report.find()
+        res.status(200).json({ message: 'All reports', data: reports });
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
-}
+};
 
-exports.getReportById = async (req, res) => {
+module.exports.getReportById = async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ message: 'Validation error', errors: errors.array() });
+    }
+
     try {
         const report = await Report.findById(req.params.id);
-        res.json(report);
+        if (!report) {
+            return res.status(404).json({ message: 'Report not found' });
+        }
+        res.status(200).json({ message: 'Report found', data: report });
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
-}
+};
 
-exports.createReport = async (req, res) => {
+module.exports.createReport = async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ message: 'Validation error', errors: errors.array() });
+    }
+
     try {
-        const newReport = new Report({
-            ...req.body,
-            book_id: req.body.bookId, 
-            user_id: 'userIdFromExternalService', 
-        });
-
-        const savedReport = await newReport.save();
-        res.status(201).json(savedReport); 
+        const newReport = await Report.create(req.body);
+        res.status(201).json({ message: 'Report created', data: newReport });
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
-}
+};
 
-exports.updateReport = async (req, res) => {
+module.exports.updateReportStatus = async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ message: 'Validation error', errors: errors.array() });
+    }
+
     try {
+        const { status } = req.body;
         const updatedReport = await Report.findByIdAndUpdate(
-            req.params.id, // Find the report by its ID
-            req.body, // Use the request body for updates
-            { new: true } // Return the updated document
+            req.params.id,
+            { status },
+            { new: true }
         );
 
         if (!updatedReport) {
             return res.status(404).json({ message: 'Report not found' });
         }
 
-        res.json(updatedReport);
+        res.status(200).json({ message: 'Report status updated', data: updatedReport });
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
-}
+};
 
-exports.deleteReport = async (req, res) => {
+module.exports.deleteReport = async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ message: 'Validation error', errors: errors.array() });
+    }
+
     try {
         const deletedReport = await Report.findByIdAndDelete(req.params.id);
 
@@ -59,8 +78,8 @@ exports.deleteReport = async (req, res) => {
             return res.status(404).json({ message: 'Report not found' });
         }
 
-        res.json(deletedReport);
+        res.status(200).json({ message: 'Report deleted' });
     } catch (error) {
         res.status(500).json({ message: error.message });
     }
-}
+};
